can answer front questions

+++ src/main/java/com/tacniz/visitormanagement/repo/DynamicQuestionRepository.java
@@ -4,6 +4,6 @@ import com.tacniz.visitormanagement.model.DynamicQuestion;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
-public interface DynamicQuestionRepository extends JpaRepository<DynamicQuestion, Integer> {
-    List<DynamicQuestion> findByVisitOptionId(Integer visitOptionId);
+public interface DynamicQuestionRepository extends JpaRepository<DynamicQuestion, Long> {
+    List<DynamicQuestion> findByVisitOptionId(Long visitOptionId);
 }
